Extract the player protocol section into a dedicated component
The output-protocol section — info note, enable/disable toggles, per-protocol configure accordion, plus its computeds and helpers — was inline in EditConfig.vue. Move it into a ProtocolConfigSection component that renders from the shared entries and emits value/action/help changes back. EditConfig.vue drops ~240 lines; no behaviour change.
